用户浏览历史记录接口优化

+++ pipayshop-api/src/main/java/com/example/pipayshopapi/service/Impl/ShopCommodityHistoryServiceImpl.java
@@ -1,6 +1,7 @@
 package com.example.pipayshopapi.service.Impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
+import com.baomidou.mybatisplus.core.conditions.update.LambdaUpdateWrapper;
 import com.example.pipayshopapi.entity.ItemCommodityHistory;
 import com.example.pipayshopapi.entity.ShopCommodityHistory;
 import com.example.pipayshopapi.entity.ShopCommodityInfo;
@@ -14,6 +15,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
+import java.util.Date;
 import java.util.List;
 
 /**
@@ -29,8 +31,9 @@ public class ShopCommodityHistoryServiceImpl extends ServiceImpl<ShopCommodityHi
 
     @Resource
     private ShopCommodityHistoryMapper shopCommodityHistoryMapper;
+
     /**
-     * 删除用户浏览网店商品的历史记录
+     * 删除用户浏览商品的历史记录
      *
      * @param userId
      * @param commodityId
@@ -41,13 +44,41 @@ public class ShopCommodityHistoryServiceImpl extends ServiceImpl<ShopCommodityHi
     public boolean deleteHistory(String userId, String commodityId) {
         if (commodityId != null) {
             return shopCommodityHistoryMapper.delete(new LambdaQueryWrapper<ShopCommodityHistory>()
-                    .eq(ShopCommodityHistory::getCommodityId, commodityId))>0;
+                    .eq(ShopCommodityHistory::getCommodityId, commodityId)
+                    .eq(ShopCommodityHistory::getUserId, userId)) > 0;
         }
         if (userId != null) {
             return shopCommodityHistoryMapper.delete(new LambdaQueryWrapper<ShopCommodityHistory>()
                     .eq(ShopCommodityHistory::getUserId, userId)) > 0;
         }
         return false;
+    }
+
+    /**
+     * 添加用户浏览商品的历史记录
+     *
+     * @param shopCommodityHistory
+     * @return
+     */
+    @Override
+    public boolean addHistory(ShopCommodityHistory shopCommodityHistory) {
+        if (shopCommodityHistory == null || shopCommodityHistory.getCommodityId() == null || shopCommodityHistory.getUserId() == null) {
+            log.error("参数不能为空");
+            return false;
+        }
+        ShopCommodityHistory histories = shopCommodityHistoryMapper.selectOne(new LambdaQueryWrapper<ShopCommodityHistory>()
+                .eq(ShopCommodityHistory::getCommodityId, shopCommodityHistory.getCommodityId())
+                .eq(ShopCommodityHistory::getUserId, shopCommodityHistory.getUserId()));
+        int result;
+        if (histories == null) {
+            result = shopCommodityHistoryMapper.insert(shopCommodityHistory);
+        } else {
+            result = shopCommodityHistoryMapper.update(null, new LambdaUpdateWrapper<ShopCommodityHistory>()
+                    .eq(ShopCommodityHistory::getCommodityId, shopCommodityHistory.getCommodityId())
+                    .eq(ShopCommodityHistory::getUserId, shopCommodityHistory.getUserId())
+                    .set(ShopCommodityHistory::getCreateTime, new Date()));
+        }
+        return result > 0;
     }
 
 }
